refactor: move 2 type def
