+++ Composer/packages/client/src/components/TriggerCreationModal/types/TriggerFormDataErrors.ts
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export interface TriggerFormDataErrors {
+  $kind?: string;
+  intent?: string;
+  event?: string;
+  triggerPhrases?: string;
+  regEx?: string;
+  activity?: string;
+}
